refactor: PaymentApprovalRequest DTO 변경 및 관련 클래스 수정
- PaymentApprovalRequest를 service 패키지에서 dto 패키지로 이동
- @NotBlank 및 @Range 유효성 검사 추가
- PaymentClient 및 TossPaymentClient 클래스 import 경로 수정
- ReservationService에 PaymentApprovalRequest import 추가
