+++ src/main/java/roomescape/reservation/config/TossPaymentClient.java
@@ -3,7 +3,7 @@ package roomescape.reservation.config;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
-import roomescape.reservation.service.PaymentApprovalRequest;
+import roomescape.reservation.dto.PaymentApprovalRequest;
 
 @Component
 public class TossPaymentClient implements PaymentClient {
