+++ src/main/java/roomescape/reservation/dto/PaymentApprovalRequest.java
@@ -1,0 +1,11 @@
+package roomescape.reservation.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import org.hibernate.validator.constraints.Range;
+
+public record PaymentApprovalRequest(
+        @NotBlank String paymentKey,
+        @NotBlank String orderId,
+        @Range(min = 1000L) long amount
+) {
+}
