--- src/main/java/roomescape/reservation/service/PaymentApprovalRequest.java
@@ -1,8 +1,0 @@
-package roomescape.reservation.service;
-
-public record PaymentApprovalRequest(
-       String paymentKey,
-       String orderId,
-       Long amount
-) {
-}
